Adds support for Il2Cpp and configuration
Adds support for Il2Cpp builds and creates a config file in the UserData folder to allow customization of rates for fiending and narcing.

Adds the option to turn off aggressive behavior when rejected or given the wrong product.
Fixes an issue where product names had "(... unpackaged)", it will now just say the product name.